Support fixture generation for classes with infix operators
Since infix operators cannot contain underscores, use a tilde instead
when generating fixture record names for infix operators.

fixes #26
